fix: revert for backward-compatibility
